Added Home/Psycho toggle to Level generator
I've adapted the generator to make a Normal World, with the cabin and guns.
You can now use the boolean value in Level to change generation between a psycho world and a normal world. Parity still broken but that should be considered 'intended' (don't do drugs!)
This mean that the mechanism for changing between levels is almost done. I now need a method to instantiate the Level prefab with the desired parameters, and then have somebody hook that up to UI.

Note, guns seem to be broken. 90% its not my fault.