ProcGen Hook Integration
Made the generation wait for a function call with 1 bool param (isHome). This should greatly facilitate building the scene-switcher, because all you need to do now is call BuildMap(bool ) when in a scene with a Level Prefab.
My work on the map then, is pretty much done.
You guys need to add in the scene-switcher and whoevers doing inventory add the loot prefabs to the psychoPrefab pool.